+++ internal/next/configmgr/config.go
@@ -1,8 +1,10 @@
 package configmgr
 
 import (
+	"fmt"
 	"net/netip"
 
+	"github.com/AdguardTeam/golibs/errors"
 	"github.com/AdguardTeam/golibs/timeutil"
 )
 
@@ -19,6 +21,38 @@ type config struct {
 	Verbose    bool `yaml:"verbose"`
 }
 
+const errNoConf errors.Error = "configuration not found"
+
+// validate returns an error if the configuration structure is invalid.
+func (c *config) validate() (err error) {
+	if c == nil {
+		return errNoConf
+	}
+
+	// TODO(a.garipov): Add more validations.
+
+	// Keep this in the same order as the fields in the config.
+	validators := []struct {
+		validate func() (err error)
+		name     string
+	}{{
+		validate: c.DNS.validate,
+		name:     "dns",
+	}, {
+		validate: c.HTTP.validate,
+		name:     "http",
+	}}
+
+	for _, v := range validators {
+		err = v.validate()
+		if err != nil {
+			return fmt.Errorf("%s: %w", v.name, err)
+		}
+	}
+
+	return nil
+}
+
 // dnsConfig is the on-disk DNS configuration.
 //
 // TODO(a.garipov): Validate.
@@ -32,6 +66,21 @@ type dnsConfig struct {
 	UseDNS64            bool              `yaml:"use_dns64"`
 }
 
+// validate returns an error if the DNS configuration structure is invalid.
+//
+// TODO(a.garipov): Add more validations.
+func (c *dnsConfig) validate() (err error) {
+	// TODO(a.garipov): Add more validations.
+	switch {
+	case c == nil:
+		return errNoConf
+	case c.UpstreamTimeout.Duration <= 0:
+		return newMustBePositiveError("upstream_timeout", c.UpstreamTimeout)
+	default:
+		return nil
+	}
+}
+
 // httpConfig is the on-disk web API configuration.
 //
 // TODO(a.garipov): Validate.
@@ -40,4 +89,18 @@ type httpConfig struct {
 	SecureAddresses []netip.AddrPort  `yaml:"secure_addresses"`
 	Timeout         timeutil.Duration `yaml:"timeout"`
 	ForceHTTPS      bool              `yaml:"force_https"`
+}
+
+// validate returns an error if the HTTP configuration structure is invalid.
+//
+// TODO(a.garipov): Add more validations.
+func (c *httpConfig) validate() (err error) {
+	switch {
+	case c == nil:
+		return errNoConf
+	case c.Timeout.Duration <= 0:
+		return newMustBePositiveError("timeout", c.Timeout)
+	default:
+		return nil
+	}
 }